requirements.txt
fastapi[all]==0.70.0
psycopg2-binary==2.9.2